Move Slate interpolation tests to Slate tests
